Refactor notification banner logic into hook
